+++ packages/components-react/library-react/src/PageBody.tsx
@@ -2,3 +2,5 @@
  * @license EUPL-1.2
  * Copyright (c) 2026 Community for NL Design System
  */
+
+export { PageBody, type PageBodyProps } from '@utrecht/page-body-react';