+++ packages/components-react/library-react/src/Toggletip.tsx
@@ -3,11 +3,9 @@
  * Copyright (c) 2026 Community for NL Design System
  */
 
-import { Alert } from '@rijkshuisstijl-community/alert-react';
-import { Icon } from '@rijkshuisstijl-community/icon-react';
 import clsx from 'clsx';
 import { DetailsHTMLAttributes, PropsWithChildren, ReactNode, Ref } from 'react';
-import { Paragraph } from './Paragraph';
+import { Alert, Icon, Paragraph } from '.';
 
 export interface ToggletipProps extends DetailsHTMLAttributes<HTMLDetailsElement> {
   ref?: Ref<HTMLDetailsElement>;