+++ packages/components-react/paragraph-react/src/index.ts
@@ -1,0 +1,3 @@
+import '@rijkshuisstijl-community/paragraph-css';
+
+export { Paragraph, type ParagraphProps } from './Paragraph';