+++ packages/storybook/src/templates/mijn-omgeving/index.tsx
@@ -1,8 +1,8 @@
 'use client'; // TODO: move to lower level at which it is actually needed, instead of wrapping the whole file
-
 import {
   AccordionProvider,
   Heading,
+  Icon,
   Link,
   MessageList,
   MessageListItem,
@@ -15,7 +15,6 @@ import {
   SideNavLink,
   SideNavList,
 } from '@rijkshuisstijl-community/components-react';
-import { Icon } from '@rijkshuisstijl-community/icon-react';
 import { PageBody } from '@utrecht/page-body-react';
 import SharedDisclaimer from '../shared/disclaimer';
 import SharedFooter from '../shared/footer';